Update RFM69_JeeLib.h
Added a definition for the AVR128DB28 DIP so breadboard/prototyping users can use this library. The user can now use #define EMONTX4 in their sketch and the compiler will use PIN_PC3 for as the chip select pin (SSpin) when it detects that PIN_PB5 is not defined. (The 28-pin part lacks a PIN_PB5). That is the only change needed to transmit JeeLib-compatible packets with the 28-pin DIP using the EMONTX4 SPI definitions.